Revert "v1.1.0-ext 배포"
