+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "recycle-study-extension",
-  "version": "1.1.0",
+  "version": "1.0.2",
   "description": "복습 URL 저장 크롬 익스텐션",
   "type": "module",
   "scripts": {
