+++ src/__tests__/api.test.js
@@ -1,14 +1,5 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import {
-  registerDevice,
-  getDevices,
-  deleteDevice,
-  saveReviewUrl,
-  getCycleOptions,
-  createCustomCycle,
-  updateCustomCycle,
-  deleteCustomCycle
-} from '../api.js';
+import { registerDevice, getDevices, deleteDevice, saveReviewUrl } from '../api.js';
 import { ERROR_CODES } from '../constants.js';
 
 // Mock chrome.runtime.sendMessage
@@ -88,14 +79,13 @@ describe('api.js', () => {
   });
 
   describe('saveReviewUrl', () => {
-    it('식별자와 주기를 포함하여 URL 저장 요청을 보낸다', async () => {
+    it('식별자를 헤더에 포함하여 URL 저장 요청을 보낸다', async () => {
       const identifier = 'my-device-id';
       const targetUrl = 'https://example.com';
-      const cycle = { type: 'DEFAULT', code: 'EBBINGHAUS' };
-      const mockResponse = { success: true, data: { url: targetUrl, scheduledAts: [] } };
+      const mockResponse = { success: true, data: { url: targetUrl } };
       sendMessageMock.mockResolvedValue(mockResponse);
 
-      await saveReviewUrl(identifier, targetUrl, cycle);
+      await saveReviewUrl(identifier, targetUrl);
 
       expect(sendMessageMock).toHaveBeenCalledWith({
         type: 'API_REQUEST',
@@ -103,120 +93,7 @@ describe('api.js', () => {
           endpoint: '/api/v1/reviews',
           method: 'POST',
           headers: { 'X-Device-Id': identifier },
-          body: { targetUrl, cycle }
-        }
-      });
-    });
-
-    it('커스텀 주기로 URL 저장 요청을 보낸다', async () => {
-      const identifier = 'my-device-id';
-      const targetUrl = 'https://example.com';
-      const cycle = { type: 'CUSTOM', id: 1 };
-      const mockResponse = { success: true, data: { url: targetUrl, scheduledAts: [] } };
-      sendMessageMock.mockResolvedValue(mockResponse);
-
-      await saveReviewUrl(identifier, targetUrl, cycle);
-
-      expect(sendMessageMock).toHaveBeenCalledWith({
-        type: 'API_REQUEST',
-        request: {
-          endpoint: '/api/v1/reviews',
-          method: 'POST',
-          headers: { 'X-Device-Id': identifier },
-          body: { targetUrl, cycle }
-        }
-      });
-    });
-  });
-
-  describe('getCycleOptions', () => {
-    it('주기 옵션 목록을 조회한다', async () => {
-      const identifier = 'my-device-id';
-      const mockResponse = {
-        success: true,
-        data: {
-          defaultOptions: [{ code: 'EBBINGHAUS', title: '에빙하우스' }],
-          customOptions: []
-        }
-      };
-      sendMessageMock.mockResolvedValue(mockResponse);
-
-      const result = await getCycleOptions(identifier);
-
-      expect(sendMessageMock).toHaveBeenCalledWith({
-        type: 'API_REQUEST',
-        request: {
-          endpoint: '/api/v1/cycles/custom',
-          method: 'GET',
-          headers: { 'X-Device-Id': identifier }
-        }
-      });
-      expect(result).toEqual(mockResponse.data);
-    });
-  });
-
-  describe('createCustomCycle', () => {
-    it('커스텀 주기를 생성한다', async () => {
-      const identifier = 'my-device-id';
-      const title = '나의 주기';
-      const durations = ['PT10M', 'PT1H', 'P1D'];
-      const mockResponse = { success: true, data: { id: 1, title, durations } };
-      sendMessageMock.mockResolvedValue(mockResponse);
-
-      const result = await createCustomCycle(identifier, title, durations);
-
-      expect(sendMessageMock).toHaveBeenCalledWith({
-        type: 'API_REQUEST',
-        request: {
-          endpoint: '/api/v1/cycles/custom',
-          method: 'POST',
-          headers: { 'X-Device-Id': identifier },
-          body: { title, durations }
-        }
-      });
-      expect(result).toEqual(mockResponse.data);
-    });
-  });
-
-  describe('updateCustomCycle', () => {
-    it('커스텀 주기를 수정한다', async () => {
-      const identifier = 'my-device-id';
-      const id = 1;
-      const title = '수정된 주기';
-      const durations = ['PT30M', 'P2D'];
-      const mockResponse = { success: true, data: { id, title, durations } };
-      sendMessageMock.mockResolvedValue(mockResponse);
-
-      const result = await updateCustomCycle(identifier, id, title, durations);
-
-      expect(sendMessageMock).toHaveBeenCalledWith({
-        type: 'API_REQUEST',
-        request: {
-          endpoint: '/api/v1/cycles/custom/1',
-          method: 'PUT',
-          headers: { 'X-Device-Id': identifier },
-          body: { title, durations }
-        }
-      });
-      expect(result).toEqual(mockResponse.data);
-    });
-  });
-
-  describe('deleteCustomCycle', () => {
-    it('커스텀 주기를 삭제한다', async () => {
-      const identifier = 'my-device-id';
-      const id = 1;
-      const mockResponse = { success: true, data: null };
-      sendMessageMock.mockResolvedValue(mockResponse);
-
-      await deleteCustomCycle(identifier, id);
-
-      expect(sendMessageMock).toHaveBeenCalledWith({
-        type: 'API_REQUEST',
-        request: {
-          endpoint: '/api/v1/cycles/custom/1',
-          method: 'DELETE',
-          headers: { 'X-Device-Id': identifier }
+          body: { targetUrl }
         }
       });
     });
