+++ src/constants.js
@@ -1,11 +1,7 @@
 /**
  * 상수 정의
  *
- * 에러 코드, UI 설정값, Duration 관련 상수를 정의한다.
- */
-
-/**
- * 에러 코드
+ * 에러 코드 및 자동 로그아웃이 필요한 에러 목록을 정의한다.
  */
 export const ERROR_CODES = {
   // 로그아웃이 필요한 에러
@@ -27,26 +23,3 @@ export const LOGOUT_REQUIRED_ERRORS = [
   ERROR_CODES.NOT_FOUND,
   ERROR_CODES.INVALID_STORAGE
 ];
-
-/**
- * UI 관련 상수
- */
-export const UI_CONSTANTS = {
-  MESSAGE_DISPLAY_DURATION_MS: 3000,      // 메시지 표시 시간 (3초)
-  DEVICE_ID_DISPLAY_LENGTH: 20            // 디바이스 ID 표시 길이
-};
-
-/**
- * Duration 관련 상수
- */
-export const DURATION_CONSTANTS = {
-  VALIDATION_STEP_MINUTES: 10,            // 분 단위 검증 기준 (10분)
-  DEFAULT_DURATION: 'PT10M',              // 기본 Duration 값
-  DEFAULT_VALUE: 10,                      // 기본 숫자 값
-  DEFAULT_UNIT: 'M'                       // 기본 단위 (분)
-};
-
-/**
- * ISO 8601 Duration 파싱용 정규식
- */
-export const DURATION_REGEX = /^P(?:(\d+)D)?(?:T(?:(\d+)H)?(?:(\d+)M)?)?$/;
