--- src/handlers/index.js
@@ -1,26 +1,0 @@
-/**
- * 핸들러 모듈 진입점
- *
- * 모든 핸들러를 re-export한다.
- */
-
-// 인증 관련
-export { handleRegister, handleCheckAuth, handleReset } from './auth.js';
-
-// 디바이스 관련
-export { handleShowDevices, handleDeleteDevice, handleLogout } from './device.js';
-
-// 주기 관련
-export {
-  handleLoadCycleOptions,
-  handleShowCycleManagement,
-  handleSaveCycle,
-  handleDeleteCycle,
-  handleEditCycle,
-  handleAddCycle,
-  handleAddDuration,
-  getCachedCycleOptions
-} from './cycle.js';
-
-// URL 저장 관련
-export { handleSaveUrl } from './url.js';
